Specify Linux constraints by default
